Add some metadata
I hope they will not break your script.
The additions should add the Title and Homepage of this the list. This will make the list look prettier in e.g. uBlock as the name is displayed:
_screenshot will be insert here soon_

Hope you like it, feel free to add the license and more information (an example is this list https://raw.githubusercontent.com/heubergen/block-eu-cookie/master/blocklist from me)